Add unsound `waiter` module
The waiter module provides types

    List and Elem.

As the README.md and the comments in the waiter.rs should make clear,
this is an **unsound** module and failure to follow the rules of its use
will lead to UB.

The waiter module pulls out the essentials of the broadcast type for
building an intrusive linked list of Waker elements that can be grown
from Future poll calls that want to park themselves in order to be woken
at a later time.

Also adds len and len_backwards methods to the linked_list.rs logic
for purposes of debugging and unit testing.